+++ app/src/main/java/org/librefit/nav/NavigationHost.kt
@@ -422,7 +422,8 @@ fun NavigationHost(
                                 )
                             },
                             doNotShowWelcomeScreenAgain =
-                                sharedViewModel::doNotShowWelcomeScreenAgain
+                                sharedViewModel::doNotShowWelcomeScreenAgain,
+                            sharedViewModel = sharedViewModel
                         )
                     }
                     entry<Route.WorkoutScreen> { route ->